fix: checkSubstring usage
when we check for substrings inside the output string, we're basically checking for multiline string,
the problem with this is that every line is a device,
so if we're looking for a device name name for a state 'running' we'll end up finding the device name
and the state 'running' of the other device.
so we should split the string into lines and check for each device and it's state separatly.